Some fixes to requests made to BAP server.
Capstone would take an instruction and address,
so we always had absolute offsets in the instruction.
BAP can give us relative offsets, so we handle that now.

Some cases still need to be fixed, and there is debug code
on right now to deal with this. This will be removed
before a pull request is made.