drop old debug message, fixup relative jumps on x86
On x86, when there is a relative jump, we follow the offset
from the end of the instruction. On ARM, we do this from
the beginning. This commit cases on that when we don't have BIL.
Hopefully when BIL support is added for x86/x86-64 this will be
handled automatically.