recursive warning, cool new runtime error checker
This commit adds an option to only check for runtime errors.
I used a corpus of test cases obtained by fuzzing libdwarf
and it seems there are many unchecked runtime errors in QIRA
static. There are more pressing things to work on than this at the
moment, but I plan to come back and make QIRA more robust with
this strategy.